requirements.txt
requests
mcp
upsonic
curl_cffi
praw
openai
tavily-python
exa_py==1.8.8
apify_client
pytest-playwright
streamlit
agno